Retire entry.py and create examples/eval.py
